extract encode code
